requirements.txt
langchain
langchain_community
langchain-astradb
sentence_transformers
datasets
transformers
pypdf
flask
python-dotenv
langchain-huggingface
accelerate
torch

-e .